removed unneeded thread tracking. Can use threading.enumerate instead when list is required

+++ actor/utils.py
@@ -17,21 +17,29 @@ ERR_MSG = 6
 LINK_MSG = 7
 UNLINK_MSG = 8
 
+
 def cleanup():
-    #create file to block any new messages being sent
-    with open(f"{FIFO_DIR}/{PID}.dwn", mode='w'):
+    # create file to block any new messages being sent
+    with open(f"{FIFO_DIR}/{PID}.dwn", mode="w"):
         pass
+
     def purge():
-        with FIFO.open(mode='rb') as f:
+        with FIFO.open(mode="rb") as f:
             _ = f.read()
+
     # do a final read, purging the file of any latent messages that may be holding processes open
     threading.Thread(target=purge, daemon=True).start()
     os.remove(FIFO)
     os.remove(f"{FIFO_DIR}/{PID}.dwn")
-    if LOG_FILE and os.path.exists(f"{LOG_FILE}/{PID}.log"):
-            os.remove(f"{LOG_FILE}/{PID}.log")
+    if (
+        LOG_FILE
+        and not os.getenv("NO_LOG_CLEANUP")
+        and os.path.exists(f"{LOG_FILE}/{PID}.log")
+    ):
+        os.remove(f"{LOG_FILE}/{PID}.log")
 
-def load_env(pid=None, log_level='INFO', log_file=None):
+
+def load_env(pid=None, log_level="INFO", log_file=None):
     import builtins
 
     builtins.FIFO_DIR = "/tmp/actor"
@@ -47,12 +55,14 @@ def load_env(pid=None, log_level='INFO', log_file=None):
         configure_logging(builtins, log_level, log_file)
     atexit.register(cleanup)
 
+
 def configure_logging(builtins, log_level, log_file):
     builtins.PROC_LOGGER = logging.getLogger("process_logger")
     builtins.PROC_LOGGER.setLevel(getattr(logging, log_level.upper()))
     fh = logging.FileHandler(f"{LOG_FILE}/{PID}.log")
+    fmt = logging.Formatter("%(asctime)s:%(levelname)s:%(message)s")
+    fh.setFormatter(fmt)
     builtins.PROC_LOGGER.addHandler(fh)
-
 
 
 def create_pipe():
@@ -62,13 +72,14 @@ def create_pipe():
     os.mkfifo(fifo)
     return fifo
 
+
 def async_msg(pid, msg, **kwargs):
     msg["r_pid"] = str(PID)
     msg["sync"] = False
     if not os.path.exists(f"{FIFO_DIR}/{pid}.dwn"):
         with open(f"{FIFO_DIR}/{pid}", "w") as fifo:
             json.dump(msg, fifo)
-    
+
 
 def sync_msg(pid, msg, **kwargs):
     async_msg(pid, msg, **kwargs)
@@ -77,8 +88,10 @@ def sync_msg(pid, msg, **kwargs):
     data["r_pid"] = actor.system.objects.Pid(data["r_pid"])
     return actor.system.objects.msg(data)
 
+
 def kill(pid):
     async_msg(pid, actor.system.objects.msg(msg_type=KILL_MSG, r_pid=PID))
+
 
 def spawn(actor_obj, log_level="info"):
     n_pid = actor.system.objects.Pid(int=uuid.uuid4().int)
